Rework of optimazation and compiliatn of calls
Added the possibility to add functions that get complex parameters but
return a noncomplex result. There is just a basic implementation of that,
further work is required.
While doing this i reworked the way function calls are compiled.
Further changes to the optimazation to calculate immediate values, so
there are less unnecessary calculations within assembler.

New function: double getReal(complex);